% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dichotomize.R
\name{dichotomize}
\alias{dichotomize}
\title{Takes a raster and returns binary representation}
\usage{
dichotomize(x, hi.val = 1, low.val = 0)
}
\arguments{
\item{x}{Input raster.}

\item{hi.val}{Value to assign to non-NA values (default = 1.0)}

\item{lo.val}{Value to assign to NA values (default = 1.0)}
}
\value{
A copy of the raster (including projection) with NA values set to \code{lo.val} 
 and non-NA values set to \code{hi.val}.
}
\description{
This simple function takes an in put raster and it returns an identical one
 that has been dichotomized (e.g., values are 0/1).
}
\examples{
library(raster)
x <- matrix( rnorm(25), nrow=5)
x[1,4] <- x[2,4] <- NA
r <- raster( x )
d <- dichotomize(r)
d
values(d)
}
\author{
Rodney J. Dyer <rjdyer@vcu.edu>
}

